+++ src/main/java/umpaz/brewinandchewin/common/utility/BnCRecipeUtils.java
@@ -1,6 +1,5 @@
 package umpaz.brewinandchewin.common.utility;
 
-import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.server.ServerLifecycleHooks;
@@ -15,10 +14,9 @@ public class BnCRecipeUtils {
     public static ItemStack getPouredItemFromFluid(FluidStack fluid) {
         if (fluid.isEmpty())
             return ItemStack.EMPTY;
-        MinecraftServer server = ServerLifecycleHooks.getCurrentServer();
-        if (server == null || !server.isDedicatedServer())
-            return BnCClientRecipeUtils.getPouredItemFromFluid(server, fluid);
-        Optional<KegPouringRecipe> recipe = server.getRecipeManager().getAllRecipesFor(BnCRecipeTypes.KEG_POURING.get()).stream().sorted(Comparator.comparing(KegPouringRecipe::isStrict)).filter(kegPouringRecipe -> kegPouringRecipe.getRawFluid().isSame(fluid.getRawFluid())).findFirst();
+        if (SideUtil.isClient())
+            return BnCClientRecipeUtils.getPouredItemFromFluid(fluid);
+        Optional<KegPouringRecipe> recipe = ServerLifecycleHooks.getCurrentServer().getRecipeManager().getAllRecipesFor(BnCRecipeTypes.KEG_POURING.get()).stream().sorted(Comparator.comparing(KegPouringRecipe::isStrict)).filter(kegPouringRecipe -> kegPouringRecipe.getRawFluid().isSame(fluid.getRawFluid())).findFirst();
         return recipe.map(KegPouringRecipe::getOutput).orElse(ItemStack.EMPTY);
     }
 }