+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## Bugfixes
-- Fixed Intoxication Hunger overlay misplacing other hud elements such as Thirst Was Taken's Thirst hud element. ([#14](https://github.com/MerchantPug/BrewinAndChewin/issues/14))
-- Fixed Tankards on Coasters having a missing texture. ([#13](https://github.com/MerchantPug/BrewinAndChewin/issues/13))
+- Fixed crashes with KubeJS when attempting to modify the mod's recipe types on a server. ([#16](https://github.com/MerchantPug/BrewinAndChewin/issues/16))